Add support for specifying yearRange
In order to change the year range when selecting a year, the yearRange needs to
be configurable. This updates the pikaday-input component to use the yearRange,
if specified. It defaults to 10, the pikaday default, if yearRange has not been
specified.

The yearRange can be a single number. For example, '10'. It sets the min and max
values - and + that value of the current year. It can also be two years,
separated by a comma. For example, '1996,2006'.

This also adds the minYear and maxYear helpers for the PikadayInteractor which are
used in the tests for yearRange for finding the min and max years.