Use dynamic event to trigger date selection
The newest Pikaday version checks if the environment it is running supports touch events. If that is the case it binds itself to touch events instead of mouse events.

PhantomJS does define touch events so Pikaday reacts accordingly and the `selectDate` helper stops working when the tests are run in the CLI.