Work around didInsertElement issue
Old code was using didInsertElement to initialize, and then called this.set().
This is no longer allowed. One recommended workaround is to use didRender, but
in our case, a stock replacement would trigger infinite loops. So I wrapped
all the initialization code into a on if() block that ensures the component
only calls the initialization code once.

It's a stopgap to get the component working again with my Ember 2.3.0 app. I'd
prefer to see a more complete refactoring that implements will DDAU and didRender
code that stores less state on the component.

But for now, the component inits properly with a passed value in Ember 2.3+.