Make setDate asychronous. (#149)
Since setMaxDate and setMinDate are asynchronoous but setDate is not, if in a
single run loop dateAnd e.g. maxDate are both changed and the new value is later
than the old value of maxDate but less than the new value,  ember-pikaday will
fail to correctly update to display the new date value.

Since I don't really understand why setMaxDate and setMinDate are being done
asynchronously, this change may have unintended consequences.