Fixes testing of multiple pikaday inputs
The removal of jquery changed the test selectors so they no longer checked for visibility. So when selecting a date from the second or greater pikaday input on a page it would not fill out the correct inputs. This has been fixed by returning the first visible element rather than the first element in the dom.